v0.2.0
Adds options for minikits required for goal and minikit bundle size.

Fixes True Jedi checks sending at unusual times.

Fixes 5-5 and 5-6 being partially swapped in the code.
This resulted in 5-5 sending 5-6 completion and vice versa, and probably
some other issues with these two levels.

Rewrites some of the connection flow to give better feedback to users.
Notably, a message is now displayed in-game when the client is fully
connected and running.