v0.2.1
Hotfix for two bugs found in the v0.2.0 release.

- Fixed Purchase Score x_ locations not sending.
- Fixed the number of minikits required to goal resetting when saving a
new game to a save file for the first time.
Reconnecting to the server would fix this issue. The client defaults to
requiring millions of minikits to goal after a reset, so this would not
have caused the goal to send by mistake.