v1.2.0 Death Link - More Goals - True Jedi Options - Ridesanity
The client should be almost compatible with all 1.x.x versions of the
apworld though it is recommended to use the latest v1.1.x client with
multiworlds generated with v1.1.x of the apworld.

What won't work when connecting the v1.2.x client to a v1.1.x slot, is
that any non power brick Extras, such as Extra Toggle and Beep Beep will
not be unlockable, because the client now expects them to be AP items
you can receive.

# Additions
## Death Link
- Separate Amnesty options for normal levels vs most top-down vehicle
levels where it can be easier to die
- Yaml option as well as a client command to toggle on/off
`/toggle_death_link`
- Option to permanently lose studs when receiving a death (it's not
possible for studs to drop on death currently), plus a further option to
have that studs loss scale with score multipliers
- Receiving Deaths can be a little weird in some cases
- In a few cases, levels that instantly respawn the player at a
checkpoint do not internally count as a death by the game, so will not
send deaths, such as in Gunship Cavalry (Original)

## More Goals
- Require completing a percentage of levels (chapters and bonus levels)
- Require collecting 7/7 Kyber Brick items from the item pool
(pretending to be like Chaos Emeralds)
- Require completing a Goal Chapter that requires completing all other
goals to unlock (in addition to the usual requirement to access the
chapter)
- Imperial Spy in 4-3 has been added to the Defeat Bosses Goal
- Minikits goal can now be set to complete automatically instead of
needing to visit the Minikits display in the Cantina's junkyard (useful
for when you have a Goal Chapter enabled)

## True Jedi Options
- Easier True Jedi option that sets True Jedi to Story requirements
- Option to scale True Jedi requirements with acquired Score
Multipliers, preventing True Jedi from becoming trivial with a few Score
Multipliers
- Note that Studs items received while in a level now contribute to True
Jedi

## Ridesanity
- Locations for riding each unique animal/turret/vehicle
- Notably adds more checks to LEGO City and New Town due to their unique
vehicles

## Non Power Brick Extras
- Extras that are unlocked in the shop from the start are now locations
to check by default, this can be disabled in the options

## Junk Items
New junk filler items have been added:
- Power Up
- Blue Stud
- Gold Stud
- Silver Stud

Studs received while inside a Free Play level will be added to your
in-level stud counts, therefore contributing to True Jedi progress

## Miscellaneous
- Option to restore High Jump height in Original trilogy levels and
other levels that cap High Jump height
- The in-game Goal progress display on the pause screen now cycles
through your goals
- The pre-unlocked hints in the Shop have been updated to include
information on the new Goal options that are enabled for your slot

## Documentation
- Linux setup guide by Flit

# Changes

## Yaml changes
- Non Power Brick Extras are shuffled by default
- `Extra Toggle` is added to the Preferred Extras option in new template
yamls
- `All Episodes Token` has been renamed to `Episode Completion Token`
- New junk items can be specified in the Junk Filler Weights

## Items
- `All Episodes Token` has been renamed to `Episode Completion Token`,
additionally, you always need 6 and will start with those that were not
added to the pool due to not having all Chapters enabled

# Fixes
- Universal Tracker support should be 100% accurate now (though be aware
that 'deferred events' support has not been added, so it will currently
assume you've beaten every level that is in logic and collected every
Gold Brick that is in logic)
- Fixed rare case where Universal Tracker's local generation could fail
if the world had only barely enough locations to fit all items
- Fixed missing Bounty Hunter or Imperial logic for 4-1 Power Brick
- Fixed missing Shortie requirement for 5-2 minikits
- Fixed missing Astromech requirement for 6-5 minikits