v1.2.1 Fix starting with too many Episode Completion Tokens
The fix affects generating with the 'All Episodes' character purchase
locations enabled, with the purchase requirements set to `tokens`. The
generator was incorrectly always starting you with 5 tokens and putting
only 1 token in the item pool.

A few parts of documentation were updated to refer to Studs generically
instead of specifically referring to the Purple Stud item.